create new file specific to status
